Refact: 광고 컴포넌트 분리
