+++ src/components/KakaoAdFit.jsx
@@ -1,51 +1,49 @@
-import React from "react";
-import styled from "styled-components";
+import React, { useState, useEffect } from 'react';
 
-export class AdfitBannerAd extends React.Component {
-    // 클래스 인스턴스를 초기화하고 초기 상태를 설정
-    constructor(props) {
-        super(props);
-    }
+const KakaoAdFit = () => {
+    const [adError, setAdError] = useState(false);  // 광고 로드 오류 상태
 
-    // 렌더링을 다시할지 여부를 결정하는 라이프사이클 메서드
-    shouldComponentUpdate() {
-        // 업데이트되지 않고 이전 렌더링 결과를 사용
-        return false;
-    }
-
-    componentDidMount() {
-        // Kakao 광고 스크립트와 인스턴스를 동적으로 생성
-        let ins = document.createElement('ins');
-        let scr = document.createElement('script');
+    useEffect(() => {
+        // 광고 스크립트 로드
+        const script = document.createElement("script");
+        const ins = document.createElement("ins");
 
         ins.className = 'kakao_ad_area';
         ins.setAttribute('style', 'display: none;');
-        scr.async = true;
-        scr.type = 'text/javascript';
-        scr.src = '//t1.daumcdn.net/kas/static/ba.min.js';
-
-        // 광고 사이즈 및 유닛 설정
+        script.async = true;
+        script.type = 'text/javascript';
+        script.src = '//t1.daumcdn.net/kas/static/ba.min.js';
+        ins.setAttribute('data-ad-unit', 'DAN-2LQytWC5DIiifh3N');
         ins.setAttribute('data-ad-width', '320');
         ins.setAttribute('data-ad-height', '100');
-        ins.setAttribute('data-ad-unit', 'DAN-2LQytWC5DIiifh3N'); // 광고 코드 부분 수정
 
-        // 광고를 삽입할 부모 요소 찾기
+        script.onload = () => {
+            console.log("광고 스크립트 로드 성공");
+            setAdError(false);
+        };
+
+        script.onerror = () => {
+            console.error("광고 스크립트 로드 실패");
+            setAdError(true);
+        };
+
         let parent = document.getElementById('adFit');
-        if (parent) {
-            parent.appendChild(ins);
-            parent.appendChild(scr);
-        }
-    }
+        parent?.appendChild(ins);
+        parent?.appendChild(script);
 
-    render() {
-        return <BannerAd id="adFit" />;
-    }
-}
+        // Clean up on component unmount
+        return () => {
+            if (parent) {
+                parent.innerHTML = '';
+            }
+        };
+    }, []);
 
-const BannerAd = styled.div`
-  width: 100%;
-  margin-top: 50px;
-  text-align: center; /* 중앙 정렬 */
-`;
+    return (
+        <div id="adFit">
+            {adError && <p>광고 로드 실패</p>}
+        </div>
+    );
+};
 
-export default AdfitBannerAd;
+export default KakaoAdFit;
